Refactor lockfile parsers
